backend/requirements.txt
bson==0.5.8
certifi==2019.11.28
dataclasses==0.6
Django==2.2.10
djangorestframework==3.11.0
djongo==1.3.1
dnspython==1.16.0
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
